+++ src/components/CategoryFilter.js
@@ -1,21 +1,25 @@
-export const catFilt = [
-  { id: 0, all: 'All' },
-  { id: 1, action: 'Action' },
-  { id: 2, biography: 'Biography' },
-  { id: 3, history: 'History' },
-  { id: 4, horror: 'Horror' },
-  { id: 5, kids: 'Kids' },
-  { id: 6, learning: 'Learning' },
-  { id: 7, sciFi: 'Sci-fi' },
-];
+import React from 'react';
+import PropTypes from 'prop-types';
+import { filterOpt } from '../utils/Options';
 
-export const filterOpt = [
-  { id: 0, name: 'All' },
-  { id: 1, name: 'Action' },
-  { id: 2, name: 'Biography' },
-  { id: 3, name: 'History' },
-  { id: 4, name: 'Horror' },
-  { id: 5, name: 'Kids' },
-  { id: 6, name: 'Learning' },
-  { id: 7, name: 'Sci-fi' },
-];
+const CategoryFilter = ({ handleFilterChange }) => {
+  const handleChange = ({ currentTarget }) => {
+    handleFilterChange(currentTarget.value);
+  };
+
+  return (
+    <select onChange={handleChange}>
+      {filterOpt.map((c) => (
+        <option key={c.id} value={c.name}>
+          {c.name}
+        </option>
+      ))}
+    </select>
+  );
+};
+
+CategoryFilter.propTypes = {
+  handleFilterChange: PropTypes.func.isRequired,
+};
+
+export default CategoryFilter;
