+++ src/index.js
@@ -6,7 +6,8 @@ import { Provider } from 'react-redux';
 import App from './components/App';
 import mainReducer from './reducers';
 import idGen from './helpers/helper';
-import { catFilt } from './components/CategoryFilter';
+import { catFilt } from './utils/Options';
+
 
 const stateInit = {
   books: [{ id: idGen(), title: 'The Da Vinci Code', category: 'Action' },
